Move these classes to download.py
