+++ spotdl/metadata.py
